게시물 리스트 페이지 수정 add버튼,삭제확인문,제목클릭시 디테일로 이동

+++ src/main/webapp/WEB-INF/jsp/article/list.jsp
@@ -3,25 +3,26 @@
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 <c:set var="pageName" value="게시물 리스트" />
 <%@ include file="../part/head.jspf"%>
-	<table>
-		<thead>
-			<tr>
-				<th>id</th>
-				<th>제목</th>
-				<th>내용</th>
-				<th>비고</th>
-			</tr>
-		</thead>
-		<tbody>
-			<c:forEach items="${list}" var="article">
+<table border="1">
+	<thead>
+		<tr>
+			<th>id</th>
+			<th>제목</th>
+			<th>내용</th>
+			<th>비고</th>
+		</tr>
+	</thead>
+	<tbody>
+		<c:forEach items="${list}" var="article">
 
-				<tr>
-					<td>${article.id }</td>
-					<td>${article.title }</td>
-					<td>${article.body }</td>
-					<td><a href="./doDelete?id=${article.id }">삭제</a></td>
-				</tr>
-			</c:forEach>
-		</tbody>
-	</table>
+			<tr>
+				<td>${article.id }</td>
+				<td><a href="./detail?id=${article.id }">${article.title }</a></td>
+				<td>${article.body }</td>
+				<td><a onclick="if(confirm('삭제하시겠습니까?') == false) return false;" href="./doDelete?id=${article.id }" >삭제</a></td>
+			</tr>
+		</c:forEach>
+	</tbody>
+</table>
+<a href="./add">게시물 추가</a>
 <%@ include file="../part/foot.jspf"%>
